test(Acceptance): 테스트 가독성을 높이기 위한 전반적인 리팩터링 - 도메인과 행동 주체에 따라 클래스 분리 (기존의 파편화되어 있던 클래스를 통합) - Nested를 사용해 시나리오별로 분류
